Change QF semantics to allow immediate values
Shift all opcodes left by four bits, making room for a four-bit
immediate value. This is currently only used in the arithmetic and seek
ops.

Note that this means values in the program stream can be negative (since
it uses signed chars), and this must be accounted for during shifts.